// app/_types/types.ts
export type ConvenioType = {
  id: number;
  nombre: string;
  descripcion: string;
};

export type HeaderPropsType = {
  titulo: string;
}

export type ConvenioGeneralesType = {
  id: number
  nombre: string
  descripcion: string
  grupo: string
  convenio: string
  subConvenio: string
  fechaRegistro: string
  moneda: string
  estado: string
  codigoFactura: string
  codigoDescuento: string
  bono: string
  unidadNegocio: string
  ejecutivoComercial: string
}